configs: imx8_var_defconfig: configure BT_HCIUART_BCM as a module
This helps to use the kernel BCM driver (CONFIG_BT_HCIUART_BCM) properly,
as it requires to load its firmware stored in the file system.

Using a module here instead will postpone the driver attachment to a point
where the file system is already up and ready to provide the firmware.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>
Co-authored-by: Pierluigi Passaro <pierluigi.p@variscite.com>