ARM: dts: imx6ul: Fix reset pins for ethernet PHYs on Variscite SOMS
- Comply whith the ethernet PHY bindings, this will also allow the
ethernet PHY reset pin to be toggled when using ifconfig ethx [down | up].

- Implement the reset timings described in KSZ8081MNX/RNB
  datasheet. Multiply by two the required timings to be safe.

"For warm reset, the reset (RST#) pin should be asserted low for a
minimum of 500 μs. The strap-in pin values are read and updated at the
de-assertion of reset. After the de-assertion of reset, wait a minimum of
100 μs before starting programming on the MIIM (MDC/MDIO) interface."

Signed-off-by: Marius Bagassien <marius.b@variscite.com>